+++ samples/MvvmSample.Core/ViewModels/Widgets/SubredditWidgetViewModel.cs
@@ -6,7 +6,6 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Threading.Tasks;
 using CommunityToolkit.Mvvm.ComponentModel;
-using CommunityToolkit.Mvvm.DependencyInjection;
 using CommunityToolkit.Mvvm.Input;
 using MvvmSample.Core.Models;
 using MvvmSample.Core.Services;
