+++ samples/MvvmSampleUwp/App.xaml.cs
@@ -57,6 +57,7 @@ sealed partial class App : Application
                 .AddTransient<MessengerPageViewModel>()
                 .AddTransient<ObservableObjectPageViewModel>()
                 .AddTransient<ObservableValidatorPageViewModel>()
+                .AddTransient<ValidationFormWidgetViewModel>()
                 .AddTransient<RelayCommandPageViewModel>()
                 .AddTransient<SamplePageViewModel>()
                 .BuildServiceProvider());
