+++ samples/MvvmSampleUwp/Properties/Default.rd.xml
@@ -33,6 +33,6 @@
     <Type Name="CommunityToolkit.Mvvm.Messaging.__Internals.__IMessengerExtensions" Dynamic="Required Public" />
       
     <!-- Validation properties -->
-    <Type Name="MvvmSample.Core.ViewModels.ObservableValidatorPageViewModel.ObservableForm" Dynamic="Required Public" />
+    <Type Name="MvvmSample.Core.ViewModels.Widgets.ValidationFormWidgetViewModel" Dynamic="Required Public" />
   </Application>
 </Directives>
