+++ samples/MvvmSampleUwp/Views/Widgets/ValidationFormWidget.xaml.cs
@@ -2,7 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using MvvmSample.Core.ViewModels;
+using CommunityToolkit.Mvvm.DependencyInjection;
+using MvvmSample.Core.ViewModels.Widgets;
 using Windows.UI.Xaml.Controls;
 
 namespace MvvmSampleUwp.Views.Widgets
@@ -12,8 +13,10 @@ namespace MvvmSampleUwp.Views.Widgets
         public ValidationFormWidget()
         {
             this.InitializeComponent();
+
+            DataContext = Ioc.Default.GetRequiredService<ValidationFormWidgetViewModel>();
         }
 
-        public ObservableValidatorPageViewModel.ObservableForm ViewModel => (ObservableValidatorPageViewModel.ObservableForm)DataContext;
+        public ValidationFormWidgetViewModel ViewModel => (ValidationFormWidgetViewModel)DataContext;
     }
 }
